backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
openai
faiss-cpu
python-dotenv
